Superclass works
