+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/BlueStageLeftLinearOpMode.java
@@ -14,7 +14,7 @@ import org.firstinspires.ftc.teamcode.util.WebCamHardware;
 import java.io.IOException;
 
 @Autonomous(name = "Blue Stage Left", group = "Group3311")
-public class BlueStageLeftLinearOpMode extends LinearOpMode
+public class BlueStageLeftLinearOpMode extends LeftRightSuper
 {
     private boolean pixelInMiddle, pixelIsLeft, pixelIsRight;
 
@@ -39,123 +39,16 @@ public class BlueStageLeftLinearOpMode extends LinearOpMode
     }
     private zone current = null;
 
-     //private enum zone
-    {
-       // center,
-        //left,
-       // right
-    }
-
     @Override
     public void runOpMode() throws InterruptedException
     {
-        try
-        {
-            blueStageRightLinearOpMode = new BlueStageRightLinearOpMode();
-            driver = new MecanumSynchronousDriver(this.hardwareMap, this);
-            webcam = new WebCamHardware(this);
-            imuControl = new ImuHardware(this);
-            initAprilTags = new InitAprilTags();
-        }
-        catch (IOException e)
-        {
-            e.printStackTrace();
-        }
-        webcam.initTfod();
+        super.runOpMode();
 
-        Recognition rec = null;
-        while((rec = webcam.findObject()) == null){
-            telemetry.addData("- Camera", "Looking for object");
-            telemetry.update();
-        }
+        planGamma(zone);
 
-        double x = (rec.getLeft() + rec.getRight()) / 2 ;
-        double y = (rec.getTop()  + rec.getBottom()) / 2 ;
-
-        LeftRightSuper.SpikeLineEnum getXPosition = webcam.findTarget(x);
-        if(getXPosition.equals("left")){
-            telemetry.addData("Left", x);
-            blueStageRightLinearOpMode.zone = BlueStageRightLinearOpMode.SpikeLineEnum.LEFT_SPIKE;
-        }else if (getXPosition.equals("center")){
-            telemetry.addData("Center", x);
-            blueStageRightLinearOpMode.zone = BlueStageRightLinearOpMode.SpikeLineEnum.CENTER_SPIKE;
-        } else if (getXPosition.equals("right")){
-            telemetry.addData("Right", x);
-            blueStageRightLinearOpMode.zone = BlueStageRightLinearOpMode.SpikeLineEnum.RIGHT_SPIKE;
-        } else telemetry.addData("OBJECT NOT DETECTED. ADJUST VALUES", "");
-//
-
-
-        switch (blueStageRightLinearOpMode.zone) {
-            case CENTER_SPIKE:
-                try {
-                    planAlpha();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                break;
-            case RIGHT_SPIKE:
-                right();
-                break;
-            case LEFT_SPIKE:
-                left();
-                break;
-            default:
-                planBeta();
-                break;
-        }
-
-        try
-        {
-            driver = new MecanumSynchronousDriver(this.hardwareMap, this);
-        }
-        catch (IOException e)
-        {
-            e.printStackTrace();
-        }
-
-        waitForStart();
-        start();
-
-//        int day = 1;
-//        switch (day)
-//        {
-//            case 1:
-//                System.out.println("It's sunday");
-//                break;
-//            case 2:
-//                System.out.println("It's Monday");
-//                break;
-//            default:
-//                System.out.println("There is a problem there are only 7 days in a week numbers is not in range");
-//                break;
-//        }
-
-
-        //Your code goes in this function.   You can make other plans as well.  (two shells are
-        //provided.
-        try {
-            planAlpha();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        //planBeta();
-
-        //Sample Test Programs
-        //aroundyTest();
-        //rotateTest();
-
-
-        while (opModeIsActive())
-        {
-
-        }
+        sleep(1000);
+        driveToTag.drive(7, zone.ordinal() + 1, 11, 0);
     }
-    private void left(){
-
-    }
-//    private void right() {
-//    }
 
     public void right()
     {
@@ -168,8 +61,6 @@ public class BlueStageLeftLinearOpMode extends LinearOpMode
         driver.forward(2,-1,0.6);
 
     }
-
-
 
     /**
      * Plan Alpha.  You will design different routes based on what intel the other team provides.
@@ -200,40 +91,14 @@ public class BlueStageLeftLinearOpMode extends LinearOpMode
 
         //Left and let AprilTag take over
         driver.rotate(-90, imuControl);
-//
-//        // Drive 25 inches forward
-//       driver.forward(25,1,0.6);
-//
-//       // Sleeps for 500 milisecond to make sure the robot is not turning
-//        sleep(500);
-//
-//        // Dives backard 5 inches
-//
-//        driver.forward(5,-1,0.6);
-//
-//        sleep(500);
-//
-//        // Turns 90 degrees left
-//        driver.turn(90, -1, .4);
-//
-//        sleep(500);
-//
-//        // After turning 90 degree drivebase drives 31 inches
-//        driver.forward(31, 1, 0.6);
-//
 
 
     }
-
-      public void pixelRight( ) {
-      }
-        public void pixelLeft( ) {
-        }
 
     /**
      * There is always a plan B.  ;)
      */
-    public void planBeta()
+    public void planGamma(SpikeLineEnum zone)
     {
 
 
@@ -241,52 +106,7 @@ public class BlueStageLeftLinearOpMode extends LinearOpMode
     }
 
 
-    /**
-     * This test rotates in place. Each step has a 3 second pause.
-     * 1.  Rotate right 90 degrees.
-     * 2.  Rotate left 90 degrees.
-     * 3.  Rotate left 180 degrees.
-     * 4.  Rotate right 360 degrees.
-     */
-    public void rotateTest()
-    {
-        double rotateSpeed = 0.5;
 
-        sleep(3000);
-        driver.turn(90, 1, rotateSpeed);
-
-        sleep(3000);
-        driver.turn(90, -1, rotateSpeed);
-
-        sleep(3000);
-        driver.turn(180, -1, rotateSpeed);
-
-        sleep(3000);
-        driver.turn(360, 1, rotateSpeed);
-    }
-
-    /**
-     * This is a sample run that drives in a "O" shape counter clockwise.
-     *
-     */
-    public void aroundyTest()
-    {
-        driver.forward(12 * 4,1,0.8);
-        sleep(100);
-        driver.turn(90, -1, .5);
-        sleep(100);
-        driver.forward(12 * 1.5,1,0.8);
-        sleep(100);
-        driver.turn(90, -1, .5);
-        sleep(100);
-        driver.forward(12 * 4,1,0.8);
-        sleep(100);
-        driver.turn(90, -1, .5);
-        sleep(100);
-        driver.forward(12 * 1.5,1,0.8);
-        sleep(100);
-        driver.turn(90, -1, .5);
-    }
 
 }
